refactor chat and child componets to useContext
