+++ components/Chat/ModelSelect.tsx
@@ -1,24 +1,30 @@
 import { IconExternalLink } from '@tabler/icons-react';
-import { FC } from 'react';
+import { useContext } from 'react';
 
 import { useTranslation } from 'next-i18next';
 
-import { OpenAIModel, OpenAIModelID } from '@/types/openai';
+import { OpenAIModel } from '@/types/openai';
 
-interface Props {
-  model: OpenAIModel;
-  models: OpenAIModel[];
-  defaultModelId: OpenAIModelID;
-  onModelChange: (model: OpenAIModel) => void;
-}
+import HomeContext from '@/pages/api/home/home.context';
 
-export const ModelSelect: FC<Props> = ({
-  model,
-  models,
-  defaultModelId,
-  onModelChange,
-}) => {
+export const ModelSelect = () => {
   const { t } = useTranslation('chat');
+
+  const {
+    state: { selectedConversation, models, defaultModelId },
+    handleUpdateConversation,
+    dispatch: homeDispatch,
+  } = useContext(HomeContext);
+
+  const handleChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
+    selectedConversation &&
+      handleUpdateConversation(selectedConversation, {
+        key: 'model',
+        value: models.find(
+          (model) => model.id === e.target.value,
+        ) as OpenAIModel,
+      });
+  };
 
   return (
     <div className="flex flex-col">
@@ -29,14 +35,8 @@ export const ModelSelect: FC<Props> = ({
         <select
           className="w-full bg-transparent p-2"
           placeholder={t('Select a model') || ''}
-          value={model?.id || defaultModelId}
-          onChange={(e) => {
-            onModelChange(
-              models.find(
-                (model) => model.id === e.target.value,
-              ) as OpenAIModel,
-            );
-          }}
+          value={selectedConversation?.model?.id || defaultModelId}
+          onChange={handleChange}
         >
           {models.map((model) => (
             <option
